requirements.txt
scapy
scikit-learn
joblib
flask
requests
python-whois
elasticsearch
pika
psycopg2-binary
geoip2
fastapi
uvicorn
